requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.18.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
huggingface-hub==0.33.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lxml==6.0.0
MarkupSafe==3.0.2
mpmath==1.3.0
narwhals==1.47.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.31.1
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyMuPDF==1.26.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rpds-py==0.26.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==5.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.47.1
streamlit==1.46.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
